Fix inconsistent js-build (#181)
* Closes #178.

Reasoning about the observations in #178, I thought this is a `yarn` issue and not something with our configs. Hence, I decided to try a yarn upgrade to `yarn berry` (yarn v3).

As we were affected by build output instabilities on some machines, I opted for the [zero-install](https://yarnpkg.com/features/zero-installs) feature of yarn, trying to get the results as deterministic as possible. This implies that we check the yarn cache into the repository. It may sound like an overkill, but it does not compare to checking-in the node-modules folder in terms of size. According to yarn's guide, it is perfectly fine.

Our inconsistency problem seems to have been fixed. I could produce stable outputs on my machine as well as the CI.

I believe developers are not affected. `yarn` should point automatically to the version in `yarn/releases`.