Cermeony ux complaints 182 (#202)
Tested on Galaxy S10+ & emulated Nexus 6 API 28
fixes #205
fixes #208
fixes 182 (partly, thats why didn't link)

* [x] rename button to "submit attestations", not "claims"
* [x] meetup counter should be a card, not button.
* [x] After a first ceremony, the community currency balance does not show automatically (needed to change community for it to update)
* [x] upon changing community, the balance changes asynchronously, so for a while, one community may show the balance value of another. This may never happen. Rather clear the balance field entirely until we can show consistent information    

In addition:  
- communities are now cached and displayed upon startup, so the user doesn't have to wait until chain connected. If he chooses another community during connection, the metadata (token and value) will be loading until connection with chain established
- metadata is also cached, so besides the actual token value, user can see all community related data straight away
- after connection to chain established, all community related data is fetched, this wasn't before and it is crucial for the correct communityMetadata to be recieved when connected to chain
- minor text changes and code cleanup
- proper navigation if start meetup pressed but then going back (before it opened the qr scanner instead of going back)

Notes: 
There appears to be an incorrect caching problem described in #219. We solve the issue of getting wrong community data by resetting store.encointer.communities, store.encointer.communityMetadata and store.encointer.chosenCid to null when changing network. Then the user chooses can choose community in the assets page

* found possible mistake in code, storage returns string "null" and not value null, so comparison should also check for string "null", i guess?

added cached communities, now communities are displayed on startup

added fetching communitiy data after network connection

flutter format . --line-length 120

communities not found refactored to dictonary and changed name instead of currency

setting communityMetadata to null is necessary for smoth transition between communities within the same chain. in setChosenCid,  webApi.encointer.getCommunityMetadata() is called which then caches again the communityMetadata. this is ok like this

In this commit, changing network results in correct community loading, but we have a caching issue of the communityMetadata.
We now don't set communityMetadata to null in the setChosenCid method. This has the advantage, that for caching we wont set it to null, but wenn we choose another community currency in the asset page, the old community currency is shown until the new one is loaded.. needs more thinking how to solve

purging communities when changing network and therefore displaying correct communities in assets page after network change

potential solution to displaying correct communities and community tokens found

formatting

displaying no community registered if chosenCid == null

removing timer, this leaves a -00:00:01 when timer done...

formatting to 120 line length

refactoring timer to card, when timer < 60, button instead of card appears.

Placing CupertinoActivityIndicator instead of empty Container. Probably the ListTile is not the best solution, but it looks nice like this.
Open for suggestions

clear balance field completely when switching between communities until consistent data is loaded

Meetup counter is card, when over "start meetup button is shown"

changing "submit claims" to "submit attestations"

* accidentally removed improvements in index.dart from other pr during rebase, reinserted them in this commit

* cleanup

* apply changes discussed in conversation of pr

* added padding

* defined caching issue.
at the moment we kind of hack a solution. it's defined in #219

* flutter format . --line-length 120

* solving merge conflicts during rebase

* compare to communities directly in ceremony tab like in assets' index.dart , because communities are cached. Else we have two different behaviours in the different tabs

* minor comment in code for better understanding

* the hint in the dropdown menu is nice for the user experience to be asked to choose community. unfortunately it breaks our integration test - can't choose community.. shold be solved in the future, for now hint is disabled

* suggestion to fix integration test.
since we reset community metadata in setChosenCid, we can't provide the metadata from the begininning. so we could supply it after resetting it, which makes more sense. best practice would be to mock the getCommunityMetadata() to return the mock

* add comment to explain code

Co-authored-by: brenzi <brenzi@users.noreply.github.com>