Fix black window on back button press in scan page (#298)
* Closes #293

Changed behavior of bottom navigation bar when scan is tapped: Instead of changing the tab, it will push the `ScanPage` route as advised by Didier. By doing this, we don't have to adjust the `ScanPage`.

Tested: Successfully pops the `ScanPage` from the `SendTx` flow and when called from the Navbar.